[2604-CHORE-42] fix: address Gemini PR69 review comments
